[user] Move onSubmit handler to separate file
